+++ app/Models/Commit.php
@@ -92,6 +92,8 @@ class Commit extends Model implements HasPresenter
                 return 'The StyleCI checks failed';
             case 3:
                 return 'The StyleCI checks have errored';
+            case 4:
+                return 'The StyleCI checks were misconfigured';
             default:
                 return 'The StyleCI checks are pending';
         }