+++ resources/views/account.blade.php
@@ -105,7 +105,7 @@
                 <div class="col-md-9">
                     <dl class="profile">
                         <dt>GitHub</dt>
-                        <dd>{{ $currentUser->username }}</dd>
+                        <dd><a href="https://github.com/{{ $currentUser->username }}" target="_blank">{{ $currentUser->username }}</a></dd>
                         <dt>Email</dt>
                         <dd>{{ $currentUser->email }}</dd>
                     </dl>