+++ resources/views/emails/failed-html.blade.php
@@ -1,6 +1,6 @@
 @extends(Config::get('core.email'))
 
 @section('content')
-<p>The coding style analysis of your commit "{{ $commit }}" on repo "{{ $repo }}" revealed problems.</p>
+<p>The coding style analysis of the commit "{{ $commit }}" on "{{ $repo }}" revealed problems.</p>
 <p>Click <a href="{{ $link }}">here</a> to see the details.</p>
 @stop