+++ resources/views/emails/failed-text.php
@@ -1,6 +1,6 @@
 Dear <?php echo e($name); ?>,
 
-The coding style analysis of your commit "<?php echo e($commit); ?>" on repo "<?php echo e($repo); ?>" revealed problems.
+The coding style analysis of the commit "<?php echo e($commit); ?>" on "<?php echo e($repo); ?>" revealed problems.
 You can see the details at <?php echo e($link); ?>.
 
 Thank you,