+++ resources/views/emails/misconfigured-text.php
@@ -1,0 +1,7 @@
+Dear <?php echo e($name); ?>,
+
+The coding style analysis of the commit "<?php echo e($commit); ?>" on "<?php echo e($repo); ?>" couldn't be completed because the repo appears to be misconfigured.
+You can see the details at <?php echo e($link); ?>.
+
+Thank you,
+StyleCI